bugfix: blank page with dynamic base path
In the last release I added `react-router`
to render the page content. Apparently
it doesn't work properly when users set a
custom "BASEPATH". I will have to rethink
it's implementation if any additional pages
are added in future versions. For now, I've
removed the router and went back to rendering
the page directly.